feat(invoices): implement invoice delivery service and related functionalities
- Add `invoice.delivery.service.ts` to handle sending, syncing, and voiding invoices with Stripe.
- Introduce `invoice.service.ts` for creating, updating, and deleting invoices, including validation and persistence logic.
- Create utility functions in `invoice.utils.ts` for calculating invoice totals and handling payout metered fees.
- Implement `invoice.webhook.service.ts` to process Stripe webhook events related to invoices, including payment received, payment failed, voided, and deleted events.
- Update invoice types in `invoices.types.ts` for consistency with new implementations.
- Modify validations in `preferences.validation.ts` and `worker-events.validation.ts` to use ISO datetime format with offsets.
- Adjust event schemas in `events-dead-letter.schema.ts` and `events.schema.ts` to ensure timestamps are stored with timezone information.
- Update email logs schema in `email-logs.schema.ts` to include timezone-aware timestamps.
- Refactor `process-stripe-webhook.ts` to use the new `invoiceWebhookService` for handling invoice-related events.
