+++ src/modules/clients/validations/client-memos.validation.ts
@@ -3,14 +3,14 @@ import { z } from '@hono/zod-openapi';
 export const createMemoSchema = z
   .object({
     content: z.string().min(1, 'Content is required'),
-    event_time: z.iso.datetime().optional(),
+    event_time: z.iso.datetime({ offset: true }).optional(),
   })
   .openapi('CreateMemo');
 
 export const updateMemoSchema = z
   .object({
     content: z.string().min(1, 'Content is required'),
-    event_time: z.iso.datetime().optional(),
+    event_time: z.iso.datetime({ offset: true }).optional(),
   })
   .openapi('UpdateMemo');
 
@@ -31,9 +31,9 @@ export const clientMemoSchema = z
     client_id: z.uuid(),
     created_by: z.uuid(),
     content: z.string(),
-    event_time: z.date().nullable(),
-    created_at: z.date(),
-    updated_at: z.date(),
+    event_time: z.iso.datetime({ offset: true }).nullable(),
+    created_at: z.iso.datetime({ offset: true }),
+    updated_at: z.iso.datetime({ offset: true }),
   })
   .openapi('ClientMemo');
