+++ src/modules/invoices/listeners.ts
@@ -10,7 +10,7 @@
  */
 
 import { getLogger } from '@logtape/logtape';
-import { loadRequiredPayoutMeteredFeeCents } from '@/modules/invoices/services/payout-metered-fee.service';
+import { loadRequiredPayoutMeteredFeeCents } from '@/modules/invoices/services/invoice.utils';
 import { METERED_TYPES } from '@/modules/subscriptions/constants/meteredProducts';
 import { meteredProductsService } from '@/modules/subscriptions/services/meteredProducts.service';
 import { db } from '@/shared/database';
