+++ src/modules/invoices/services/invoice.webhook.service.ts
@@ -2,15 +2,11 @@ import { getLogger } from '@logtape/logtape';
 import type { Stripe } from 'stripe';
 import { invoicesRepository } from '@/modules/invoices/database/queries/invoices.repository';
 import { db } from '@/shared/database';
-import { InvoicePaymentFailed, InvoiceVoided, InvoiceDeleted } from '@/shared/events/definitions';
+import { InvoiceDeleted, InvoicePaymentFailed, InvoiceVoided } from '@/shared/events/definitions';
 import { InvoiceStripePaymentReceived } from '@/modules/invoices/types/events';
 
-const logger = getLogger(['invoices', 'webhooks-service']);
+const logger = getLogger(['invoices', 'webhook-service']);
 
-/**
- * Handle invoice.paid event from Stripe webhook
- * Thin handler: store event + queue async processing
- */
 const handleInvoicePaid = async (stripeInvoice: Stripe.Invoice): Promise<void> => {
   const invoice = await invoicesRepository.findInvoiceByStripeId(stripeInvoice.id);
   if (!invoice) {
@@ -18,13 +14,6 @@ const handleInvoicePaid = async (stripeInvoice: Stripe.Invoice): Promise<void> =
     return;
   }
 
-  logger.info('Storing invoice.paid event for async processing: {invoiceId}', {
-    invoiceId: invoice.id,
-    stripeInvoiceId: stripeInvoice.id,
-    amount: stripeInvoice.amount_paid,
-  });
-
-  // Store to outbox with critical flag (persist before response)
   await InvoiceStripePaymentReceived.dispatch(
     {
       invoice_id: invoice.id,
@@ -48,28 +37,17 @@ const handleInvoicePaid = async (stripeInvoice: Stripe.Invoice): Promise<void> =
       critical: true,
     }
   );
-
-  logger.info('Invoice.paid event queued for async processing');
 };
 
-/**
- * Handle invoice.payment_failed event
- */
 const handleInvoicePaymentFailed = async (stripeInvoice: Stripe.Invoice): Promise<void> => {
   const invoice = await invoicesRepository.findInvoiceByStripeId(stripeInvoice.id);
   if (!invoice) {
+    logger.warn('Invoice not found for Stripe ID: {stripeInvoiceId}', { stripeInvoiceId: stripeInvoice.id });
     return;
   }
 
   await db.transaction(async (tx) => {
-    await invoicesRepository.updateInvoice(
-      invoice.id,
-      invoice.organization_id,
-      {
-        status: 'overdue',
-      },
-      tx
-    );
+    await invoicesRepository.updateInvoice(invoice.id, invoice.organization_id, { status: 'overdue' }, tx);
 
     await InvoicePaymentFailed.dispatch(
       {
@@ -85,28 +63,17 @@ const handleInvoicePaymentFailed = async (stripeInvoice: Stripe.Invoice): Promis
       }
     );
   });
-
-  logger.info('Payment failed for invoice {invoiceId}', { invoiceId: invoice.id });
 };
 
-/**
- * Handle invoice.voided event
- */
 const handleInvoiceVoided = async (stripeInvoice: Stripe.Invoice): Promise<void> => {
   const invoice = await invoicesRepository.findInvoiceByStripeId(stripeInvoice.id);
   if (!invoice) {
+    logger.warn('Invoice not found for Stripe ID: {stripeInvoiceId}', { stripeInvoiceId: stripeInvoice.id });
     return;
   }
 
   await db.transaction(async (tx) => {
-    await invoicesRepository.updateInvoice(
-      invoice.id,
-      invoice.organization_id,
-      {
-        status: 'cancelled',
-      },
-      tx
-    );
+    await invoicesRepository.updateInvoice(invoice.id, invoice.organization_id, { status: 'cancelled' }, tx);
 
     await InvoiceVoided.dispatch(
       {
@@ -123,16 +90,12 @@ const handleInvoiceVoided = async (stripeInvoice: Stripe.Invoice): Promise<void>
       }
     );
   });
-
-  logger.info('Invoice {invoiceId} voided', { invoiceId: invoice.id });
 };
 
-/**
- * Handle invoice.deleted event (for draft invoices)
- */
 const handleInvoiceDeleted = async (stripeInvoice: Stripe.Invoice): Promise<void> => {
   const invoice = await invoicesRepository.findInvoiceByStripeId(stripeInvoice.id);
   if (!invoice) {
+    logger.warn('Invoice not found for Stripe ID: {stripeInvoiceId}', { stripeInvoiceId: stripeInvoice.id });
     return;
   }
 
@@ -153,34 +116,20 @@ const handleInvoiceDeleted = async (stripeInvoice: Stripe.Invoice): Promise<void
       }
     );
   });
-
-  logger.info('Invoice {invoiceId} deleted via Stripe', { invoiceId: invoice.id });
 };
 
-/**
- * Type guard for Stripe Invoice
- */
 const isStripeInvoice = (obj: unknown): obj is Stripe.Invoice =>
   obj !== null && typeof obj === 'object' && 'object' in obj && obj.object === 'invoice';
 
-/**
- * Process a Stripe invoice event
- */
 const processEvent = async (event: Stripe.Event): Promise<void> => {
-  const eventType = event.type;
   const stripeInvoice = event.data.object;
 
   if (!isStripeInvoice(stripeInvoice)) {
-    logger.warn('Received Stripe event without invoice object: {eventType}', { eventType });
+    logger.warn('Received Stripe event without invoice object: {eventType}', { eventType: event.type });
     return;
   }
 
-  logger.info('Processing invoice webhook event {eventType} for Stripe Invoice {stripeInvoiceId}', {
-    eventType,
-    stripeInvoiceId: stripeInvoice.id,
-  });
-
-  switch (eventType) {
+  switch (event.type) {
     case 'invoice.paid':
       await handleInvoicePaid(stripeInvoice);
       break;
@@ -194,19 +143,14 @@ const processEvent = async (event: Stripe.Event): Promise<void> => {
       await handleInvoiceDeleted(stripeInvoice);
       break;
     default:
-      logger.info('Unhandled invoice event type: {eventType}', { eventType });
+      logger.info('Unhandled invoice event type: {eventType}', { eventType: event.type });
   }
 };
 
-/**
- * Invoice Webhooks Service
- *
- * Processes Stripe invoice-related webhook events.
- */
-export const invoiceWebhooksService = {
+export const invoiceWebhookService = {
   processEvent,
   handleInvoicePaid,
   handleInvoicePaymentFailed,
   handleInvoiceVoided,
   handleInvoiceDeleted,
-};
+} as const;
