+++ src/modules/invoices/types/invoices.types.ts
@@ -35,7 +35,7 @@ export type InvoiceWithRelations = BuildQueryResult<
   Schema['invoices'],
   {
     with: {
-      lineItems: true;
+      line_items: true;
       client: {
         with: { user: true };
       };
@@ -77,8 +77,6 @@ export interface InvoiceListResponse {
 export type CreateInvoiceRequest = z.infer<typeof invoiceValidations.createInvoiceSchema>;
 export type UpdateInvoiceRequest = z.infer<typeof invoiceValidations.updateInvoiceSchema>;
 export type ListInvoicesQuery = z.infer<typeof invoiceValidations.listInvoicesQuerySchema>;
-export type InvoiceResponse = z.infer<typeof invoiceValidations.invoiceSchema>;
-export type InvoiceSummaryResponse = z.infer<typeof invoiceValidations.invoiceSummarySchema>;
 export type InvoiceLineItemResponse = z.infer<typeof invoiceValidations.lineItemSchema>;
 
 /**
